+++ src/main/java/heavysnow/heath/controller/MemberController.java
@@ -5,6 +5,8 @@ import heavysnow.heath.dto.*;
 import heavysnow.heath.dto.goal.GoalCreationDto;
 import heavysnow.heath.dto.goal.GoalIdResponseDto;
 import heavysnow.heath.dto.goal.GoalUpdateDto;
+import heavysnow.heath.dto.member.MemberDto;
+import heavysnow.heath.dto.member.MemberResponseDto;
 import heavysnow.heath.dto.post.PostListResponse;
 import heavysnow.heath.exception.UnauthorizedException;
 import heavysnow.heath.service.GoalService;
