+++ src/main/java/heavysnow/heath/dto/member/MemberGoalInfo.java
@@ -1,4 +1,4 @@
-package heavysnow.heath.dto;
+package heavysnow.heath.dto.member;
 
 
 import heavysnow.heath.domain.Goal;
