+++ src/test/java/heavysnow/heath/service/LikedServiceTest.java
@@ -1,7 +1,7 @@
 package heavysnow.heath.service;
 
 import heavysnow.heath.domain.MemberPostLiked;
-import heavysnow.heath.dto.MemberDto;
+import heavysnow.heath.dto.member.MemberDto;
 import heavysnow.heath.dto.post.PostAddRequest;
 import heavysnow.heath.dto.post.PostDetailResponse;
 import heavysnow.heath.repository.MemberPostLikedRepository;
