+++ src/test/java/heavysnow/heath/service/PostServiceTest.java
@@ -1,7 +1,7 @@
 package heavysnow.heath.service;
 
 import heavysnow.heath.dto.comment.CommentCreateDto;
-import heavysnow.heath.dto.MemberDto;
+import heavysnow.heath.dto.member.MemberDto;
 import heavysnow.heath.dto.post.PostAddRequest;
 import heavysnow.heath.dto.post.PostEditRequest;
 import heavysnow.heath.dto.post.PostDetailResponse;
